layout major: introducing 6-button chorded keyboard
this is, obviously a major change and I have to relearn everything :(

but the good news:

- with 6 buttons, we have 6 single keys combinatins and 9 double-key
  combinations. that gives us many more letters.

- we should still be able to touch-type because the thumb should rest in
  the middle button, and move either up or down.

- we don't need to have any "delayed release" combos yet

This will obviously make typing letters faster because the combos will
be less complex. I'm not entirely sure what that means in terms on "not
having to look while you type" but the benefits still outruns this
doubt.

note that the layer is still incomplete and needs things like
punctuation, numbers and modifier keys.