layout major: 6 buttons => 10 buttons
It's not fun having to re-learn all these layouts all the time. But I
concluded that with 3 buttons on each thumb, the "delayed press" and
"delayed release" combinations are quite time consuming.

There's a magic balance between few (and large!) buttons, and the number
of combinations that you'll need to be able to type fast. On the one
side of the spectrum, you've got the 4-button keyboard that this project
started with. You barely have to move your thumb and you can't miss the
giant 2 buttons that each thumb is supposed to aim at. On the other
hand, the "delayed right press, delayed left release" combinations make
typing quite slow.

On the other side of the spectrum, you've got the full QWERTY
keyboard. The "combinations" are superfast in that you press a single
key for any given letter. However, it comes with a price: you miss keys
_a lot_. That's why we have dictionaries (and also because we can't
spell..?).

Anyhow, I think a 10-button keypad might work. I now have enough
combinations to spell out all the 26 letters with a single thumb tap or
single tap with two thumbs. The buttons are still quite large and from
my first day of experimentation, it seems this will be a good
compromise.

The layout looks like this:

   .   .
   .. ..
   .. ..

Also, it's not difficult to add another two buttons on the top. But
these are a little trickier to reach with your thumb. We'll see if we
need them! So far so good, but I haven't added numbers and all the
symbols.