layout break: moving P,B,V and Q
I set aside the middle-right button on the right hand (aka "space") in
combination with backspace, enter and friends. now with the new swipe
gestures, these don't seem to be used all that much. the swipe backspace
is more used as far as I can tell.

the good news is that I only have to relearn a couple of letters! and
the "Q" was a little uncomfortable to reach anyway. let's see how this
turns out!