refac major: data representation sequence -> stroke
the data prepresentation is now no longer a sequence of "x.... .....",
but an "image" like this:

00000-00000 00000-00010
00000-00000 00000-00010
00000-00000 00000-01000

this is simpler. each button gets 5 counters: one for taps, and 1 for
each swipe in the directions left, up, down, right (in that order). no
more than 9 repetitions. this way, the order of how you perform a swipe
or tap doesn't matter, only the end result.

also, keys like DPAD_RIGHT is now no longer possible (you need to
release all fingers in order for a character to appear). this is less
flexible, but it simplifies the internals _a lot_. we don't even need a
60ms postDelayed anymore!

this should also make it easier to label the keys. also, I made the 10
buttons from v1 into a 4x3 grid. so we have 12 buttons now, but the
layout can stay as-is, it works pretty well (plus, I spent 2 months
learning it).

let's see where this takes us!