ui: remove labels when not touching
because it's ugly. also, make it more transparent so you can barely see
it. this is less newbie-friendly but more expert-friendly.

I know what my layout is now, I don't need to see it. It makes the
background harder to see and it's ugly. As a matter of fact, I'd love if
I could remove the UI completely!