layout: handle strokes in ThumbkeyboardIME
now the ThumbkeyboardView just asks a handler to handle a stroke.

TODO: allow the ThumbkeyboardView to ask for layout/something so that
it can show the labels (they are pretty much a must for usability).

note that this is just a skelaton, layout.get(..) will always return
null since the HashMap is currently always empty (fix coming up).